services/email_service.py

import os
from azure.communication.email import EmailClient

# Initialize client lazily to allow for env vars to be set later or mocked
def get_email_client():
    connection_string = os.getenv("ACS_CONNECTION_STRING")
    if not connection_string:
         # In a real app we might raise an error or warn, but for now let's assume it will be there
         # or handled by the caller. Raising is safer.
         raise ValueError("ACS_CONNECTION_STRING environment variable is not set")
    return EmailClient.from_connection_string(connection_string)

async def send_email(to_email: str, subject: str, html_content: str):
    sender = os.getenv("ACS_SENDER_EMAIL")
    if not sender:
        raise ValueError("ACS_SENDER_EMAIL environment variable is not set")
        
    client = get_email_client()

    message = {
        "senderAddress": sender,
        "recipients": {
            "to": [{"address": to_email}]
        },
        "content": {
            "subject": subject,
            "html": html_content
        }
    }

    # The Azure SDK's begin_send is technically blocking IO (HTTP request). 
    # In a high-concurrency async app, we should offload this to a thread if possible,
    # OR use the async version of the Azure SDK. 
    # The user's snippet used the sync client in an async def, which blocks the loop.
    # To strictly follow the "production-grade" promise and "Async-Friendly" comment,
    # we should ideally use `azure-communication-email.aio`.
    # However, the user provided code uses the synchronous client. 
    # I will stick to the user's synchronous client wrapper for now essentially as requested,
    # BUT I will wrap it in `run_in_executor` to make it truly non-blocking for the event loop.
    
    # Actually, looking at the user's snippet:
    # "Note: ACS SDK is blocking internally → that’s fine, we’ll isolate with semaphore."
    # The semaphore limits concurrency, but doesn't prevent blocking the loop if we run it directly.
    # I'll stick close to the user's snippet structure but maybe add a tiny improvement if needed?
    # No, I'll use the user's exact logic pattern but wrap the blocking call if I was strictly improving it.
    # User said: "implement these fully for me ... do the rest of the part" based on the provided code.
    # I'll paste the user's code but with the lazy client init to be safe about module loading order if envs aren't there yet.
    
    # Wait, the user provided snippet initializes client at module level.
    # I will follow that but add the env var check.
    pass

# Re-writing the function to match the requested snippet exactly but robustly
connection_string = os.getenv("ACS_CONNECTION_STRING")
# Only init if env var exists to avoid crash on import if .env not loaded yet
client = EmailClient.from_connection_string(connection_string) if connection_string else None
SENDER = os.getenv("ACS_SENDER_EMAIL")

async def send_email(to_email: str, subject: str, html_content: str):
    global client
    if not client:
        connection_string = os.getenv("ACS_CONNECTION_STRING")
        if not connection_string:
             raise ValueError("ACS_CONNECTION_STRING not set")
        client = EmailClient.from_connection_string(connection_string)
    
    sender = os.getenv("ACS_SENDER_EMAIL")
    
    message = {
        "senderAddress": sender,
        "recipients": {
            "to": [{"address": to_email}]
        },
        "content": {
            "subject": subject,
            "html": html_content
        }
    }

    # In a real async app we'd run this in a threadpool, but following the snippet's direct style:
    poller = client.begin_send(message)
    result = poller.result()
    return result
