pyproject.toml
[project]
name = "email-sender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-communication-email>=1.1.0",
    "fastapi>=0.128.0",
    "fastexcel>=0.18.0",
    "jinja2>=3.1.6",
    "openpyxl>=3.1.5",
    "polars>=1.37.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "uvicorn>=0.40.0",
]
